pyproject.toml
[project]
name = "ttf2stitch"
version = "0.1.0"
description = "Convert TTF/OTF cross-stitch fonts to bitmap JSON v2 for Stitchx"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "vdirienzo" }]
dependencies = [
    "fonttools>=4.50.0",
    "pillow>=10.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
]

[project.scripts]
ttf2stitch = "ttf2stitch.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ttf2stitch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]
